requirements.txt
aiohttp==3.7.4.post0
aiohttp-apispec==2.2.1
aiohttp-pydantic==1.12.0
aiomisc==15.2.16
apispec==3.3.2
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
colorlog==6.5.0
greenlet==1.1.2
idna==3.3
Jinja2==2.11.3
MarkupSafe==2.0.1
marshmallow==3.14.0
multidict==5.2.0
pydantic==1.8.2
python-dotenv==0.19.1
requests==2.26.0
SQLAlchemy==1.4.26
swagger-ui-bundle==0.0.9
typing-extensions==3.10.0.2
urllib3==1.26.7
webargs==5.5.3
yarl==1.7.0
